+++ src/server/devices/device-list.ts
@@ -1,4 +1,4 @@
-import { API, OwnSonosDevice, OwnSonosDevices } from '../services';
+import { AcDevice, API, OwnDaikinDevice, OwnDaikinDevices, OwnSonosDevice, OwnSonosDevices } from '../services';
 import { IBaseDevice } from './baseDeviceInterfaces';
 
 export class DeviceList {
@@ -10,8 +10,8 @@ export class DeviceList {
     return this._ids;
   }
 
-  public getDevices(): Array<IBaseDevice | OwnSonosDevice> {
-    const result: Array<IBaseDevice | OwnSonosDevice> = [];
+  public getDevices(): Array<IBaseDevice | OwnSonosDevice | AcDevice> {
+    const result: Array<IBaseDevice | OwnSonosDevice | AcDevice> = [];
 
     for (const dID of this._ids) {
       const d = API.getDevice(dID);
@@ -20,9 +20,13 @@ export class DeviceList {
       }
     }
     for (const name of this._ids) {
-      const s = OwnSonosDevices.ownDevices[name];
+      const s: OwnSonosDevice = OwnSonosDevices.ownDevices[name];
       if (s !== undefined) {
         result.push(OwnSonosDevices.ownDevices[name]);
+      }
+      const daikinDevice: OwnDaikinDevice = OwnDaikinDevices.ownDevices[name];
+      if (daikinDevice !== undefined) {
+        result.push(OwnDaikinDevices.ownDevices[name]);
       }
     }
     return result;