+++ src/server/services/ac/index.ts
@@ -1,3 +1,4 @@
 export * from './daikin-service';
 export * from './own-daikin-device';
 export * from './own-daikin-devices';
+export { AcDevice } from './ac-device';