+++ CHANGELOG.md
@@ -4,6 +4,12 @@
   Placeholder for the next version (at the beginning of the line):
   ## **WORK IN PROGRESS**
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Ac Device as superclass from OwnDaikinDevice
+* (theimo1221) Add Roomsetting for default button callbacks
+
 ## 0.1.42-alpha.3 (2022-07-12)
 
 * (theimo1221) Try reconnect unreachable Daikin Devices